Move the treatment of traffic light position related calculation to waypoint_tracker.py
